+++ src/map/packets/CMakeLists.txt
@@ -33,7 +33,5 @@ set(PACKET_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/server_message.h
     ${CMAKE_CURRENT_SOURCE_DIR}/objective_utility.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/objective_utility.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/zone_in.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/zone_in.h
     PARENT_SCOPE
 )
