+++ src/map/packets/c2s/0x00a_login.cpp
@@ -20,11 +20,11 @@
 */
 
 #include "0x00a_login.h"
+#include "packets/s2c/0x00a_login.h"
 
 #include "entities/charentity.h"
 #include "packets/s2c/0x008_enterzone.h"
 #include "packets/s2c/0x04f_equip_clear.h"
-#include "packets/zone_in.h"
 #include "utils/charutils.h"
 #include "utils/gardenutils.h"
 #include "utils/zoneutils.h"
@@ -123,7 +123,7 @@ void GP_CLI_COMMAND_LOGIN::process(MapSession* PSession, CCharEntity* PChar) con
     if (PChar->loc.zone != nullptr)
     {
         PChar->pushPacket<GP_SERV_COMMAND_EQUIP_CLEAR>();
-        PChar->pushPacket<CZoneInPacket>(PChar, PChar->currentEvent);
+        PChar->pushPacket<GP_SERV_COMMAND_LOGIN>(PChar, PChar->currentEvent);
         PChar->pushPacket<GP_SERV_COMMAND_ENTERZONE>(PChar);
     }
 }
